Filter-out double hits.
Such scenario is possible when searching for two words that are similar,
according to lexdist. One has also be a subset of another, for example
"linux" and "linuxa". This would generate two hits for two different
words at the same location.

There's still a problem with words that are at max hit position.